fix(custom-elements-lsp): convert analyzer script to typescript FUI-1331
