+++ packages/core/custom-elements-lsp/src/parser/analyze.types.ts
@@ -1,0 +1,11 @@
+export type PluginConfig = {
+  srcRouteFromTSServer?: string;
+  designSystemPrefix?: string;
+  parser?: {
+    timeout?: number;
+    fastEnable?: boolean;
+    src?: string;
+    dependencies?: string[];
+  };
+  plugins?: string[];
+};
