+++ packages/core/custom-elements-lsp/src/parser/modules.d.ts
@@ -1,0 +1,1 @@
+declare module 'resolve/sync.js';
