--- packages/core/custom-elements-lsp/src/parser/package.json
@@ -1,3 +1,0 @@
-{
-	"type": "module"
-}
